// frontend/src/app/api/schema.d.ts
export interface components {
    schemas: {
// ...
        CategoryDto: {
            /** Format: int32 */
            id?: number;
            name?: string | null;
            description?: string | null;
            imageUrl?: string | null;
            /** Format: int32 */
            parentCategoryId?: number | null;
            type?: string | null;
            productsId?: number[] | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}